refactor: reorganize phpunit package into Configuration/ and TestRunner/ directories
Group 10+ root-level files into logical subdirectories matching the extension package's organization pattern. Update all import paths and barrel exports.
