+++ packages/extension/package.json
@@ -4,7 +4,7 @@
     "displayName": "PHPUnit & Pest Test Explorer",
     "icon": "img/icon.png",
     "publisher": "recca0120",
-    "version": "3.9.15",
+    "version": "3.9.16",
     "private": true,
     "license": "MIT",
     "repository": {
