+++ packages/phpunit/src/Configuration/BinaryDetector.ts
