+++ packages/phpunit/src/Configuration/TestGlobPattern.ts
