+++ packages/phpunit/src/Configuration/index.ts
@@ -1,0 +1,6 @@
+export * from './BinaryDetector';
+export * from './Configuration';
+export * from './PathReplacer';
+export * from './PHPUnitXML';
+export * from './TestGlobPattern';
+export * from './XmlElement';
