+++ packages/phpunit/src/ProcessBuilder/ProcessBuilder.ts
@@ -2,9 +2,8 @@ import type { SpawnOptions } from 'node:child_process';
 import { randomBytes } from 'node:crypto';
 import { mkdir } from 'node:fs/promises';
 import { join } from 'node:path';
-import type { IConfiguration } from '../Configuration';
+import type { IConfiguration, PathReplacer } from '../Configuration';
 import { CMD_TEMPLATE, CMD_TEMPLATE_QUOTED } from '../constants';
-import type { PathReplacer } from '../PathReplacer';
 import type { TestResult } from '../TestOutput';
 import { cloneInstance, parseArgv } from '../utils';
 import { base64DecodeFilter, base64EncodeFilter } from './FilterEncoder';
