+++ packages/phpunit/src/TestRunner/TestRunnerObserver.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from 'vitest';
-import { TeamcityEvent, type TestFailed } from './TestOutput';
+import { TeamcityEvent, type TestFailed } from '../TestOutput';
 import {
     createTestRunnerEventProxy,
     TestRunnerEvent,
