+++ packages/phpunit/src/index.ts
@@ -1,7 +1,4 @@
-export * from './BinaryDetector';
 export * from './Configuration';
-export * from './PathReplacer';
-export * from './PHPUnitXML';
 export * from './ProcessBuilder';
 export * from './TestCollection';
 export * from './TestCoverage';
@@ -10,8 +7,5 @@ export * from './TestOutput';
 export * from './TestParser';
 export { initTreeSitter, resolveWasmDir } from './TestParser/tree-sitter/TreeSitterParser';
 export * from './TestRunner';
-export * from './TestRunnerObserver';
-export * from './TestRunnerProcess';
 export * from './types';
 export * from './utils';
-export * from './XmlElement';
